sched/tune: Reduce the number of boost groups to 7
Currently, 8 boost groups for schedtune is unnecessary and
setting it to 7 doesn't warn about it being less than what
it's supposed to be at dmesg. However setting it to 6 warns
with the following log:

[    6.131933] Trying to create more than 6 SchedTune boosting groups

Therefore, set schedtune boost groups to 7 as our ROM doesn't
need 8 boost groups for schedtune at the moment.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>